Changed persistence to default to hsqldb.
The default setup for the local database is now a file-based hsqldb instance. This allows testing the application with persisted data but without having to create a database instance in some pre-installed server.

The MySQL configuration is still available, but needs to be configured using the local/coralwatch.properties override file.