Added support for longitude/latitude in DMS form.
The field for longitude and latitude now allow entering the data in four different forms:

- decimal with sign
- decimal with trailing E/W
- degree/minute/second (DMS) like "-152d23m12s"
- DMS with trailing E/W

The validation is done only on the client side, the conversion on the server side -- not the best solution, but the quickest. The information is stored as float and always rendered in decimal form. We might want to support both formats later.