Implemented first version of tab-based longLat entries.
The form for surveys now has a tab for entering the longitude and latitude either in decimal or DMS form, the fields get updated automatically whenever the opposite ones change.

This is not well tested yet, has display issues when the form gets loaded (I fear we need a "Loading..." message) and it might need some adjustments on the server side. But it has to do for now since it is late :-)