Changed setup of application in web.xml.
Previously we gave the component to the Restlet servlet, but that causes Restlet to listen on the port configured since the component defined a HTTP server. By giving the application to the servlet this issue is avoided.

I also moved the main method to start the component into the component's class. This way the application class can actually be used without the component (in the WAR scenario), but we can still start the component if a standalone setup is wanted.