Added new flag for a "test mode" which allows resetting the database.
This is intended for black-box testing (making it slightly less black). If the properties file contains a property "testMode" set to a true value (acoording to Boolean.valueOf(..)), then an additional resource is bound to "/reset", which causes the whole database to be emptied and the two default users to be recreated.

It allows the black-box testing to have independent test cases. I added quite a few checks for the test mode flag to make sure this feature is not accidentally accessible.