Added a bit of a sleep after clicks to avoid race conditions with Dojo.
It seemed that sometimes we try to click things before the JavaScript code on the page was fully executed. I'm not entirely sure if that is correct and I don't know if sleeping 200ms is going to fix it consistently, but I don't have the time to solve the problem properly. This might fix the problem at least on the build server -- we'll see.