Changed the waitingForDojo bits a bit.
After a playing some more and reading things on Google I got the impression that the problem is really just relating to the clickButton(..) part where Dojo seems to go off into the background to do something about getting a new page.

I restricted the waiting to the clickButton(..) calls now and expanded the waiting to a full second for good measure. All wild guessing, but we see how it will work.