Added infrastructure for uploading Excel files to extend/replace the database.
Nothing really done yet, but the basic structure is in place. Upload/replace is only available if the application is run in test mode -- while it could be used as a backup scheme, it is not intended to be used on a production system. If necessary the production system could be set into test mode while offline anyway.