Got basic import of data for migration working.
It is possible to upload the spreadsheet with the old database now. A lot of issues left, mostly due to mismatches of the old model and the new. In particular it is not really clear how to group survey records into surveys -- same location/date/time seems ok at first, but the weather field can change over these. The current version of the code just creates one survey per record, which is not good.