Added consistency checks and some better handling of missing values.
If an email address or country name is missing on one entry, but provided later, then it will be added.

If two different values are provided for either it is considered an error.

If longitude or latitude can not be determined, they will not be set to -9999 anymore, but left unset.