Another attempt at finding a good way of importing the old data.
This time we create new users and reefs only if we have a new survey, but we do create multiple anonymous users/unknown reefs (using numbers). This avoids combining everything into one, which might make it easier to fix the data and --more importantly -- means that we don't get one big reef spanning lots of surveys. Some of these nameless reefs even have a country set, in which case the data was lost previously.