Added two new plot types.
We now have a basic pie plot showing the distribution of the coral shapes and a bar chart for the rounded down averages of the color numbers. They both still lack styling (just the JFreeChart defaults at the moment).

I extracted the chart generation code into a "service" class, which is really just a module holding statics. Not pretty, but will do for now.